# platform-api/plugins/eventgateway/openapi.yaml
openapi: "3.0.3"
info:
  title: Event Gateway Plugin API
  version: "0.9"
  description: >
    WebSub and WebBroker API management endpoints provided by the
    eventgateway compile-time plugin.
paths:
...
  /api/v0.9/websub-apis/{apiId}/api-keys:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
    post:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:api_key:create
            - ap:websub_api:api_key:manage
      responses:
        '201':
          description: WebSub API key created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v0.9/websub-apis/{apiId}/api-keys/{keyName}:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
      - name: keyName
        in: path
        required: true
        schema:
          type: string
    put:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:api_key:update
            - ap:websub_api:api_key:manage
      responses:
        '200':
          description: WebSub API key updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:api_key:delete
            - ap:websub_api:api_key:manage
      responses:
        '204':
          description: WebSub API key deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v0.9/webbroker-apis/{apiId}/api-keys:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
    post:
      security:
        - OAuth2SecurityEventGateway:
            - ap:webbroker_api:api_key:create
            - ap:webbroker_api:api_key:manage
      responses:
        '201':
          description: WebBroker API key created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v0.9/webbroker-apis/{apiId}/api-keys/{keyName}:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
      - name: keyName
        in: path
        required: true
        schema:
          type: string
    put:
      security:
        - OAuth2SecurityEventGateway:
            - ap:webbroker_api:api_key:update
            - ap:webbroker_api:api_key:manage
      responses:
        '200':
          description: WebBroker API key updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      security:
        - OAuth2SecurityEventGateway:
            - ap:webbroker_api:api_key:delete
            - ap:webbroker_api:api_key:manage
      responses:
        '204':
          description: WebBroker API key deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
...
  responses:
    BadRequest:
      description: Bad request
    Unauthorized:
      description: Unauthorized
    NotFound:
      description: Not found
...
    InternalServerError:
      description: Internal server error
    ServiceUnavailable:
      description: Service unavailable
  securitySchemes:
    OAuth2SecurityEventGateway:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth2/token
          scopes: {}